requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anndata==0.8.0
annoy==1.17.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
bbknn==1.5.1
beautifulsoup4==4.12.3
bleach==5.0.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.16.0
charset-normalizer==2.1.1
chex==0.1.3
commonmark==0.9.1
contourpy==1.2.0
cycler==0.12.1
Cython==0.29.32
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.7
docrep==0.3.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
faiss-gpu==1.7.2
fastjsonschema==2.15.3
flatbuffers==2.0
flax==0.5.0
fonttools==4.50.0
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
get-annotations==0.1.2
glob2==0.7
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
h5py==3.10.0
hdbscan==0.8.33
idna==3.6
igraph==0.11.4
importlib_metadata==7.0.2
importlib_resources==6.3.2
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jax==0.4.13
jaxlib==0.4.13
jedi==0.18.1
Jinja2==3.1.3
joblib==1.3.2
jsonpickle==2.2.0
jsonschema==4.5.1
jupyter-client==7.3.4
jupyter_core==5.0.0
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
KDEpy==1.1.0
kiwisolver==1.4.5
leidenalg==0.10.2
llvmlite==0.39.1
Markdown==3.3.7
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.3
mistune==0.8.4
ml-dtypes==0.3.2
msgpack==1.0.3
multidict==6.0.2
multipledispatch==0.6.0
munkres==1.1.4
natsort==8.4.0
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.2.1
notebook==6.4.11
numba==0.56.4
numpy==1.21.6
numpyro==0.9.2
oauthlib==3.2.0
openpyxl==3.0.10
opt-einsum==3.3.0
optax==0.1.2
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
pyg-lib @ file:///media/hiccup/3b31548a-917a-47ff-83b7-93a0e647802c/downloads/pyg_lib-0.2.0%2Bpt111cu113-cp39-cp39-linux_x86_64.whl#sha256=e6c833dcffca365323bd50dbd443198f1b61ea811e91e6fb440498cfbf26f0ac
Pygments==2.17.2
pynndescent==0.5.11
pyparsing==3.1.2
pyro-api==0.1.2
pyro-ppl==1.8.1
pyrsistent==0.18.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.5.10
pytz==2024.1
pyvis==0.3.1
PyYAML==5.4.1
pyzmq==23.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==12.4.4
rsa==4.8
scanpy==1.9.8
scikit-learn==1.4.1.post1
scikit-misc==0.1.4
scipy==1.11.4
scvi-tools==0.16.0
seaborn==0.13.2
Send2Trash==1.8.0
session-info==1.0.0
SIMLR==0.1.3
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.14.1
stdlib-list==0.10.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==1.1.0
terminado==0.15.0
texttable==1.7.0
threadpoolctl==3.4.0
tinycss2==1.1.1
toolz==0.11.2
torch==1.11.0+cu113
torch-cluster @ file:///media/hiccup/3b31548a-917a-47ff-83b7-93a0e647802c/downloads/torch_cluster-1.6.0-cp39-cp39-linux_x86_64.whl#sha256=ebb133c0c89963674d635501c269e11d217fee3453c79fe301f6d22003b1b297
-e git+https://github.com/Hzzone/torch_clustering.git@b7fb417f7a144a8332ed9f9417e212f2a6ea7c05#egg=torch_clustering
torch-scatter @ file:///media/hiccup/3b31548a-917a-47ff-83b7-93a0e647802c/downloads/torch_scatter-2.0.9-cp39-cp39-linux_x86_64.whl#sha256=1c9b2666062c20704941f26f52afebdb624f8cc40e665d564f34d54ad912a4ed
torch-sparse @ file:///media/hiccup/3b31548a-917a-47ff-83b7-93a0e647802c/downloads/torch_sparse-0.6.15-cp39-cp39-linux_x86_64.whl#sha256=70a6f5a174609566acd5c0d5d7877c273feea01941f9c0d238bf12a7cb0130e0
torch-spline-conv @ file:///media/hiccup/3b31548a-917a-47ff-83b7-93a0e647802c/downloads/torch_spline_conv-1.2.1-cp39-cp39-linux_x86_64.whl#sha256=6b36c0c7de836122a001c951aecb089491ddaeda62b0a5d3d180cf70837e967c
torch_geometric==2.0.4
torchaudio==0.11.0+cu113
torchmetrics==0.9.0
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.66.2
traitlets==5.2.2.post1
typing_extensions==4.2.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
wget==3.2
widgetsnbextension==3.6.0
xlrd==1.2.0
yarl==1.7.2
zipp==3.18.1
